#802 restructure timeline module tests
