+++ src/a07_calendar_logic/test/test_/test_config_creg_.py
@@ -4,12 +4,6 @@ from src.a03_group_logic.group import awardlink_shop
 from src.a04_reason_logic.reason_concept import reasonunit_shop
 from src.a05_concept_logic.concept import conceptunit_shop
 from src.a06_plan_logic.plan import planunit_shop
-from src.a07_calendar_logic.chrono import (
-    get_c400_constants,
-    get_min_from_dt,
-    get_timeline_min_difference,
-    get_year_rope,
-)
 from src.a07_calendar_logic.test._util.a07_str import (
     c100_str,
     c400_clean_str,
@@ -47,6 +41,12 @@ from src.a07_calendar_logic.test._util.calendar_examples import (
     get_thu,
     get_tue,
     get_wed,
+)
+from src.a07_calendar_logic.timeline import (
+    get_c400_constants,
+    get_min_from_dt,
+    get_timeline_min_difference,
+    get_year_rope,
 )
