+++ src/a07_calendar_logic/test/test_/test_year_calc_func.py
@@ -1,8 +1,8 @@
-from src.a07_calendar_logic.chrono import get_first_weekday_index_of_year
 from src.a07_calendar_logic.test._util.calendar_examples import (
     get_creg_config,
     get_five_config,
 )
+from src.a07_calendar_logic.timeline import get_first_weekday_index_of_year
 
 
 def test_get_first_weekday_index_of_year_ReturnsObj_Scenario0_creg_timeline_config():
